requirements.txt
# Helix Factor Strategy - Dependencies
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
fredapi>=0.5.0
wandb>=0.16.0
optuna>=3.0.0
tqdm>=4.65.0
mlflow>=2.0.0
